Extract TeaserImage component, refactor tests
